shared Worldcat,SFTP,Platform,Solr fixtures
